requirements.txt
opencv-python
mediapipe
pyautogui
numpy
filterpy
SpeechRecognition
PyAudio
pystray
Pillow
pandas
scikit-learn
pyinstaller

